Remove debug statements
The source of the deadlock has been found.  The issue is that the queue fills
up quickly and we soon have exhausted all 10 messages, which is the queue's
maximum.  Successive calls to mq_send then block indefinitely.  This limit of
10 is enforced by ulimit so there is no way to extend the size (at least from
within ff).  That means I have to look for an alternative queue :(